backend/app/services/ocr_stub.py
"""
OCR service stub using pdfminer.six for text extraction from native PDFs.

TODO: Replace stub with production OCR service (Tesseract, AWS Textract, Google Vision, etc.)
TODO: Add support for scanned PDFs (currently only works with native/digital PDFs)
TODO: Implement proper bounding box extraction from PDF layout
TODO: Add OCR confidence scoring
"""
import logging
from typing import Dict, List, Any
from pathlib import Path
from pdfminer.high_level import extract_pages
from pdfminer.layout import LTTextContainer, LTChar, LAParams

logger = logging.getLogger(__name__)


def process_pdf(file_path: str) -> Dict[str, Any]:
    """
    Extract text and layout information from a PDF file.
    
    This is a STUB implementation using pdfminer.six for native PDFs.
    It does NOT perform actual OCR on scanned documents.
    
    Args:
        file_path: Path to PDF file
        
    Returns:
        Dictionary with structure:
        {
            'pages': [
                {
                    'page_no': int,
                    'text': str,
                    'bbox': dict or None,  # Bounding box info
                    'confidence': float or None  # OCR confidence (stub returns None)
                },
                ...
            ]
        }
        
    Raises:
        Exception: If PDF cannot be processed
        
    TODO: Implement real OCR for scanned documents
    TODO: Extract detailed bounding boxes for text elements
    TODO: Add image extraction and processing
    TODO: Handle encrypted/password-protected PDFs
    """
    logger.info(f"Processing PDF (stub): {file_path}")
    
    try:
        pages_data = []
        
        # Use pdfminer to extract text from native PDFs
        for page_num, page_layout in enumerate(extract_pages(file_path, laparams=LAParams()), start=1):
            page_text = []
            page_bbox = {
                'x0': page_layout.bbox[0],
                'y0': page_layout.bbox[1],
                'x1': page_layout.bbox[2],
                'y1': page_layout.bbox[3]
            }
            
            # Extract text from all text containers
            for element in page_layout:
                if isinstance(element, LTTextContainer):
                    text = element.get_text()
                    if text.strip():
                        page_text.append(text)
            
            combined_text = '\n'.join(page_text).strip()
            
            pages_data.append({
                'page_no': page_num,
                'text': combined_text,
                'bbox': page_bbox if combined_text else None,
                'confidence': None  # Stub: no confidence for native PDF text
            })
            
            logger.debug(f"Extracted {len(combined_text)} chars from page {page_num}")
        
        result = {
            'pages': pages_data,
            'metadata': {
                'page_count': len(pages_data),
                'ocr_engine': 'pdfminer.six (stub)',
                'is_native_pdf': True
            }
        }
        
        logger.info(f"PDF processing complete: {len(pages_data)} pages extracted")
        return result
        
    except Exception as e:
        logger.error(f"Failed to process PDF {file_path}: {e}", exc_info=True)
        raise


def extract_text_with_tesseract(image_path: str) -> Dict[str, Any]:
    """
    Placeholder for Tesseract OCR integration.
    
    TODO: Implement Tesseract OCR for scanned documents
    TODO: Install pytesseract and configure Tesseract path
    TODO: Add preprocessing (deskew, denoise, binarization)
    TODO: Extract word-level bounding boxes and confidence
    
    Args:
        image_path: Path to image file or page image from PDF
        
    Returns:
        OCR results with text, bounding boxes, and confidence
    """
    raise NotImplementedError("Tesseract OCR integration not yet implemented")


def extract_text_with_textract(file_path: str) -> Dict[str, Any]:
    """
    Placeholder for AWS Textract integration.
    
    TODO: Implement AWS Textract for production OCR
    TODO: Configure AWS credentials and region
    TODO: Handle async job submission for large documents
    TODO: Parse Textract response format (blocks, lines, words)
    TODO: Extract tables and forms if needed
    
    Args:
        file_path: Path to PDF file
        
    Returns:
        OCR results from Textract
    """
    raise NotImplementedError("AWS Textract integration not yet implemented")
