backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6
rq==1.15.1
redis==5.0.1
pdfminer.six==20221105
sentence-transformers==2.2.2
numpy==1.26.2
watchdog==3.0.0
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
